refactor Profile page with separate Tab components
